+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - feat(protocol): added unix socket supports.
 - feat(protocol): accepts socket factory function to produce sockets.
+- feat(connection): added ended/finished properties and finished event.
 - feat(encoder): accepts string as data to send.
 - feat(connector): added `IConnection.writable` property.